Added file "factory" and command "move"
"move" is just a stub, and will be used to move between rooms in the game
world.

"factory" is a list of objects that need to be generated in the game, such
as items and rooms. It will be expaneded later with things such as
optional names and descriptions as parameters, etc.